Remove zlib-sync
This is an optional dependency of discord.js, providing "WebSocket data
compression and inflation". I don't know how much benefit it gives, but
to install it requires Python, and I'm trying to get this package to
install easily on Railway.app, which does not have Python with its
default configuration.