Handle multiple images being posted at once a little better
If in recount mode, ignore it (no change).

If in live mode and awaiting a match, two images posted at once is
treated as just a tag match, not a tag match and new tag, but the user
is warned in case they intended it to be a tag match and new tag.

If in live mode and awaiting a new tag, two images posted at once is
treated as a single new tag, and it's assumed to be two views of the
same location, perhaps as an extra clue. No warning is posted.